# contracts/phase3/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "Assessment": {
        "additionalProperties": false,
        "properties": {
          "assessment_id": {
            "format": "uuid",
            "title": "Assessment Id",
            "type": "string"
          },
          "assessment_status": {
            "$ref": "#/components/schemas/AssessmentStatus"
          },
          "diagnosis_status": {
            "default": "unknown",
            "enum": [
              "suspected",
              "supported",
              "unknown"
            ],
            "title": "Diagnosis Status",
            "type": "string"
          },
          "error_family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Family"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "first_error_step": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Error Step"
          },
          "misconception_candidate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Misconception Candidate"
          },
          "primary_skill_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Skill Id"
          },
          "reason_code": {
            "enum": [
              "valid_solution",
              "invalid_transformation",
              "wrong_final_answer",
              "valid_partial",
              "ambiguous_input",
              "unsupported_scope",
              "cannot_verify"
            ],
            "title": "Reason Code",
            "type": "string"
          },
          "validator_version": {
            "title": "Validator Version",
            "type": "string"
          },
          "verification_method": {
            "enum": [
              "symbolic_and_rule",
              "reference",
              "none"
            ],
            "title": "Verification Method",
            "type": "string"
          }
        },
        "required": [
          "assessment_id",
          "assessment_status",
          "reason_code",
          "verification_method",
          "validator_version"
        ],
        "title": "Assessment",
        "type": "object"
      },
      "AssessmentStatus": {
        "enum": [
          "verified_correct",
          "verified_incorrect",
          "unverified",
          "needs_clarification"
        ],
        "title": "AssessmentStatus",
        "type": "string"
      },
      "CreateSession": {
        "additionalProperties": false,
        "properties": {
          "demo_profile_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "title": "Demo Profile Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "topic_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "title": "Topic Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "demo_profile_id",
          "topic_id"
        ],
        "title": "CreateSession",
        "type": "object"
      },
      "DemoLogin": {
        "additionalProperties": false,
        "properties": {
          "demo_profile_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "title": "Demo Profile Id",
            "type": "string"
          }
        },
        "required": [
          "demo_profile_id"
        ],
        "title": "DemoLogin",
        "type": "object"
      },
...
      "ProblemPublic": {
        "additionalProperties": false,
        "properties": {
          "concept_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "title": "Concept Id",
            "type": "string"
          },
          "content_version": {
            "title": "Content Version",
            "type": "string"
          },
          "difficulty": {
            "enum": [
              "introductory",
              "standard"
            ],
            "title": "Difficulty",
            "type": "string"
          },
          "domain": {
            "const": "real_numbers",
            "default": "real_numbers",
            "title": "Domain",
            "type": "string"
          },
          "equation": {
            "title": "Equation",
            "type": "string"
          },
          "family_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "title": "Family Id",
            "type": "string"
          },
          "primary_skill_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "title": "Primary Skill Id",
            "type": "string"
          },
          "problem_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "title": "Problem Id",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "problem_id",
          "content_version",
          "concept_id",
          "primary_skill_id",
          "family_id",
          "prompt",
          "equation",
          "difficulty"
        ],
        "title": "ProblemPublic",
        "type": "object"
      },
...
      "SessionView": {
        "additionalProperties": false,
        "properties": {
          "available_next": {
            "title": "Available Next",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_opportunity_id": {
            "format": "uuid",
            "title": "Current Opportunity Id",
            "type": "string"
          },
          "draft": {
            "default": "",
            "maxLength": 2000,
            "title": "Draft",
            "type": "string"
          },
          "eligible_observations": {
            "title": "Eligible Observations",
            "type": "integer"
          },
          "hint_level": {
            "default": 0,
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Hint Level",
            "type": "integer"
          },
          "mastery": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Mastery",
            "type": "array"
          },
          "mode": {
            "const": "internal_demo",
            "default": "internal_demo",
            "title": "Mode",
            "type": "string"
          },
          "problem": {
            "$ref": "#/components/schemas/ProblemPublic"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "review_status": {
            "title": "Review Status",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "state_version": {
            "minimum": 1.0,
            "title": "State Version",
            "type": "integer"
          },
          "status": {
            "enum": [
              "active",
              "paused",
              "completed"
            ],
            "title": "Status",
            "type": "string"
          },
          "student_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
            "title": "Student Id",
            "type": "string"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/TurnPublic"
            },
            "title": "Turns",
            "type": "array"
          },
          "tutoring_enabled": {
            "default": false,
            "title": "Tutoring Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "session_id",
          "student_id",
          "status",
          "state_version",
          "current_opportunity_id",
          "problem",
          "created_at",
          "review_status",
          "relation",
          "turns",
          "eligible_observations",
          "available_next"
        ],
        "title": "SessionView",
        "type": "object"
      },
      "TurnPublic": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "submit",
              "request_hint"
            ],
            "title": "Action",
            "type": "string"
          },
          "assessment": {
            "$ref": "#/components/schemas/Assessment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "opportunity_id": {
            "format": "uuid",
            "title": "Opportunity Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "response": {
            "additionalProperties": true,
            "title": "Response",
            "type": "object"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "title": "Steps",
            "type": "array"
          },
          "turn_id": {
            "format": "uuid",
            "title": "Turn Id",
            "type": "string"
          }
        },
        "required": [
          "turn_id",
          "opportunity_id",
          "request_id",
          "action",
          "steps",
          "assessment",
          "response",
          "created_at"
        ],
        "title": "TurnPublic",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Mở — Socratic tutor API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/demo/login": {
      "post": {
        "operationId": "login_api_v1_demo_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Login"
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "operationId": "history_api_v1_sessions_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "History"
      },
      "post": {
        "operationId": "create_api_v1_sessions_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create"
      }
    },
...
  }
}